search contexts: API for converting version contexts (#19250)
Adds API support for converting version contexts to search contexts (for site admins) and support for searching with converted search contexts.

This also makes using search contexts possible on non-sourcegraph.com instances (a feature flag still has to be enabled to use the search context dropdown, but it's possible directly through the context: filter).

+++ internal/types/types.go
@@ -1482,7 +1482,7 @@ type Secret struct {
 }
 
 type SearchContext struct {
-	ID int32
+	ID int64
 	// Name contains the non-prefixed part of the search context spec.
 	// The name is a substring of the spec and it should NOT be used as the spec itself.
 	// The spec contains additional information (such as the @ prefix and the context namespace)
@@ -1490,8 +1490,13 @@ type SearchContext struct {
 	// Example mappings from context spec to context name:
 	// global -> global, @user -> user, @org -> org,
 	// @user/ctx1 -> ctx1, @org/ctx2 -> ctx2.
-	Name            string
-	Description     string
+	Name        string
+	Description string
+	// Public property controls the visibility of the search context. Public search context is available to
+	// any user on the instance. If a public search context contains private repositories, those are filtered out
+	// for unauthorized users. Private search contexts are only available to their owners. Private user search context
+	// is available only to the user, private org search context is available only to the members of the org, and private
+	// instance-level search contexts is available only to site-admins.
 	Public          bool
 	NamespaceUserID int32 // if non-zero, the owner is this user. NamespaceUserID/NamespaceOrgID are mutually exclusive.
 	NamespaceOrgID  int32 // if non-zero, the owner is this organization. NamespaceUserID/NamespaceOrgID are mutually exclusive.
